# docs/notificaciones-openapi.yaml
openapi: 3.1.0

info:
  title: Notificaciones y cupones — Vitahub Pro y Mi Vitahub
  version: "1.2.0"
  description: |
    Dispara una notificación push a uno o varios profesionales de Vitahub Pro.

    La notificación llega al teléfono (si el profesional tiene la app instalada)
    y SIEMPRE queda guardada en su panel dentro de la plataforma web, así que
    también la ve quien entra desde la computadora.

    Pensada para automatizaciones: una plataforma externa escucha un evento y
    pega acá. Por eso la respuesta informa a cuántos llegó realmente, y los
    reintentos no duplican mientras se mande `event_id`.

    Además, el agente puede regalarle a un profesional un cupón de Shopify de un
    solo uso (por monto o por porcentaje): le aparece en su wallet y le llega un
    aviso al teléfono. Mismo secreto, mismas reglas de reintento.

    Desde la 1.2.0 también se puede notificar a los CLIENTES de la app Mi Vitahub
    (`/api/webhooks/notificar-clientes`): ofertas, novedades, avisos. Cada envío
    queda como campaña y se puede consultar cuántos la recibieron, leyeron y abrieron.

servers:
  - url: https://pro.vitahub.mx
    description: Producción

security:
  - bearerAuth: []

paths:
  /api/webhooks/notificar:
    post:
      summary: Enviar una notificación push
      operationId: enviarNotificacion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Notificacion"
            examples:
              aUnProfesionalPorCorreo:
                summary: A un profesional, identificado por su correo
                value:
                  title: Tu paciente agendó una consulta
                  body: Ana Pérez reservó para el jueves a las 10:00
                  target: paola@ejemplo.com
                  url: https://pro.vitahub.mx/mis-citas
                  evento: booking.creado
                  event_id: bk_8821
              aVarios:
                summary: A varios a la vez, mezclando correos e identificadores
                value:
                  title: Cambio en el catálogo
                  body: Tres productos de tus protocolos cambiaron de presentación
                  target: ["ana@ejemplo.com", 8203251581249]
                  event_id: cat_2026_10_05
              aTodos:
                summary: A toda la base de profesionales
                value:
                  title: Nueva función disponible
                  body: Ya puedes cobrar desde el armador de protocolos
                  target: all
                  url: https://pro.vitahub.mx/armador-carritos
                  event_id: release_1_3_0
      responses:
        "200":
          description: |
            Procesado. Puede ser un envío real (`Resumen`) o un evento que ya se
            había notificado antes y se descartó (`Duplicado`).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Resumen"
                  - $ref: "#/components/schemas/Duplicado"
              examples:
                enviado:
                  summary: Envío real
                  value:
                    ok: true
                    destinatarios: 12
                    enviados: 9
                    sin_token: 3
                    fallidos: 0
                    errores: []
                duplicado:
                  summary: Mismo event_id que un envío anterior
                  value:
                    ok: true
                    duplicado: true
                    mensaje: Ese evento ya se había notificado
        "400":
          description: |
            Falta `title` o `body`, el cuerpo no es JSON válido, o `url` no
            empieza con http:// ni https://
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                ok: false
                error: title y body son requeridos
        "401":
          description: Falta el header Authorization o el secreto no coincide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                ok: false
                error: No autorizado
        "404":
          description: |
            Ningún correo o identificador de `target` corresponde a un
            profesional registrado. El mensaje nombra cuáles fallaron.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                ok: false
                error: "No hay afiliados con esos correos: nadie@ejemplo.com"
        "500":
          description: |
            Error interno. Reintentar es seguro SIEMPRE QUE se haya mandado
            `event_id`: el reintento se descarta como duplicado en vez de
            mandar la notificación dos veces.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/webhooks/notificar-clientes:
    post:
      summary: Enviar una notificación a clientes de Mi Vitahub
      description: |
        Llega como push a quien tenga la app con notificaciones activas y SIEMPRE
        queda en su historial dentro de la app. Cada llamada crea una campaña.
      operationId: notificarClientes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificacionCliente"
            examples:
              oferta:
                summary: Oferta a todos los clientes
                value:
                  title: 20% en Omega 3 este fin de semana
                  body: Aprovecha y no interrumpas tu protocolo. Válido hasta el domingo.
                  target: all
                  url: https://vitahub.mx/collections/omega-3
                  evento: oferta-omega3
                  event_id: oferta_omega3_2026_10_12
              pacientesDeUnProfesional:
                summary: A los pacientes de un profesional
                value:
                  title: Tu profesional tiene nuevos horarios
                  body: Ya puedes agendar tu consulta de seguimiento desde la app.
                  target: { especialista: 8203251581249 }
                  event_id: horarios_8203251581249_oct
              porCorreo:
                summary: A uno o varios clientes por correo
                value:
                  title: Tu pedido especial ya está listo
                  body: Pasa por él o pídelo con envío desde vitahub.mx
                  target: ["ana@ejemplo.com", "luis@ejemplo.com"]
      responses:
        "200":
          description: Enviado (`ResumenCliente`) o descartado por `event_id` repetido (`Duplicado`).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ResumenCliente"
                  - $ref: "#/components/schemas/Duplicado"
        "400":
          description: Falta `title`/`body`, `url` no es https, o `target` no se reconoce
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Secreto inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Ningún cliente corresponde al `target` (incluye `faltan` con los correos sin cuenta)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: Métricas de una campaña a clientes
      operationId: metricasCampanaClientes
      security:
        - bearerAuth: []
      parameters:
        - name: event_id
          in: query
          schema: { type: string }
          description: El `event_id` con que se mandó (o usar `campaign_id`)
        - name: campaign_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Métricas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstadisticasCampana"
        "404":
          description: No existe esa campaña
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/webhooks/cupon:
    post:
      summary: Crear un cupón de un solo uso y asignarlo a un profesional
      operationId: crearCupon
      description: |
        Crea el cupón en Shopify (de un solo uso, combinable con otros descuentos)
        y lo deja en la wallet del profesional, junto con un aviso a su teléfono.

        Un cupón por llamada y por persona. Para regalarle a varios, llamar una vez
        por profesional, cada una con su propio `event_id`.

        Topes: monto de $1 a $2,000; porcentaje entero de 1 a 30.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Cupon"
            examples:
              porMonto:
                summary: $200 de crédito
                value:
                  target: paola@ejemplo.com
                  tipo: monto
                  valor: 200
                  nota: Por tu reseña del producto
                  event_id: resena_5521
              porPorcentaje:
                summary: 10% de descuento, vence en 30 días
                value:
                  target: paola@ejemplo.com
                  tipo: porcentaje
                  valor: 10
                  vence_en_dias: 30
                  nota: Bienvenida
                  event_id: bienvenida_paola
                  title: ¡Felicidades! 🎉
                  body: Ganaste un {descuento} de bienvenida. Usa el código {codigo} en tu próxima compra.
      responses:
        "201":
          description: Cupón creado y asignado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CuponCreado"
        "200":
          description: |
            Ese `event_id` ya había generado un cupón: no se creó otro, se devuelve
            el código del primero.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CuponDuplicado"
        "400":
          description: |
            Falta o es inválido `tipo`, `valor`, `target` o `vence_en_dias`, o el valor
            pasa el tope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Falta el header Authorization o el secreto no coincide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No hay un profesional con ese correo o identificador.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: |
            Error interno. Reintentar es seguro SIEMPRE QUE se haya mandado
            `event_id`. Si el error dice que el cupón se creó pero no se pudo asignar,
            la respuesta trae el `code`: avisar al equipo, no reintentar sin `event_id`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Secreto compartido, entregado por el equipo de Vitahub.
        Header: `Authorization: Bearer <secreto>`.
        Se compara en tiempo constante: probar valores parciales no da ninguna
        pista sobre el secreto correcto.

  schemas:
    NotificacionCliente:
      type: object
      required: [title, body]
      properties:
        title: { type: string, maxLength: 65, description: Título del push }
        body: { type: string, maxLength: 240, description: Texto del push }
        target:
          description: |
            A quién. Por defecto "all".
              · "all"        todos los clientes con cuenta (sin push, solo historial)
              · "con_app"    solo los que tienen notificaciones activas
              · correo o lista de correos
              · { especialista: <shopify_customer_id del profesional> }  sus pacientes
          oneOf:
            - type: string
            - type: array
              items: { type: string, format: email }
            - type: object
              properties:
                especialista: { type: integer }
        url: { type: string, format: uri, description: Se abre al tocar la notificación (solo https) }
        evento: { type: string, maxLength: 60, description: Etiqueta libre para agrupar campañas }
        event_id: { type: string, maxLength: 120, description: Idempotencia; reintentos con el mismo id no duplican }
    ResumenCliente:
      type: object
      properties:
        ok: { type: boolean }
        campaign_id: { type: string, format: uuid }
        destinatarios: { type: integer }
        enviados: { type: integer, description: Push aceptados por Expo }
        sin_token: { type: integer, description: Sin notificaciones activas; lo ven solo en su historial }
        fallidos: { type: integer }
        errores: { type: array, items: { type: string } }
        faltan: { type: array, items: { type: string }, description: Correos sin cuenta en la app }
    EstadisticasCampana:
      type: object
      properties:
        ok: { type: boolean }
        id: { type: string, format: uuid }
        title: { type: string }
        evento: { type: string }
        event_id: { type: string }
        created_at: { type: string, format: date-time }
        destinatarios: { type: integer }
        enviados: { type: integer }
        sin_token: { type: integer }
        fallidos: { type: integer }
        leidas: { type: integer, description: La vieron en su historial }
        abiertas: { type: integer, description: La tocaron (push o historial) }
        tasa_apertura: { type: number, description: abiertas / destinatarios }
    Notificacion:
      type: object
      required: [title, body]
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          description: Título de la notificación.
          example: Tu paciente agendó una consulta
        body:
          type: string
          minLength: 1
          description: Cuerpo del mensaje.
          example: Ana Pérez reservó para el jueves a las 10:00
        target:
          description: |
            A quién le llega. Si se omite, va a TODOS los profesionales con la
            app instalada.

            Acepta cuatro formas:
              - "all"                        → todos
              - "paola@ejemplo.com"          → por correo (lo habitual)
              - 8203251581249                → por shopify_customer_id
              - ["ana@ejemplo.com", 820325…] → lista, mezclando ambos

            Usar el correo es lo normal: la plataforma de origen suele conocer
            al profesional por su correo y no por su identificador interno.
          default: all
          oneOf:
            - type: string
              description: '"all", un correo, o un shopify_customer_id'
            - type: integer
              format: int64
              description: shopify_customer_id
            - type: array
              minItems: 1
              items:
                oneOf:
                  - type: string
                  - type: integer
                    format: int64
          examples:
            - all
            - paola@ejemplo.com
            - 8203251581249
            - ["ana@ejemplo.com", 8203251581249]
        url:
          type: string
          format: uri
          pattern: "^https?://"
          description: |
            Adónde lleva la notificación al tocarla. Solo http:// o https://;
            cualquier otro esquema se rechaza con 400.

            Puede apuntar a una ficha de producto dentro de la plataforma, con
            el identificador de producto: el profesional la abre con su
            comisión y su dosificación a la vista, en lugar de caer en la
            tienda pública.
          example: https://pro.vitahub.mx/armador-carritos?producto=9606654722369
        evento:
          type: string
          maxLength: 60
          description: |
            Nombre del evento de origen. Queda guardado junto a la
            notificación, para poder rastrear después de dónde salió.
          example: booking.creado
        event_id:
          type: string
          maxLength: 120
          description: |
            Identificador único del evento. MANDARLO SIEMPRE.

            Las plataformas reintentan ante un timeout o un 500, y sin esto la
            misma notificación le llega dos veces al teléfono del profesional.
            Con esto, el segundo intento responde {"ok":true,"duplicado":true}
            y no envía nada.

            Conviene usar el identificador que el evento ya tiene en la
            plataforma de origen: el id de la reserva, del pago, del ticket.
          example: bk_8821

    Cupon:
      type: object
      required: [target, tipo, valor]
      additionalProperties: false
      properties:
        target:
          description: |
            El profesional que recibe el cupón: su correo o su shopify_customer_id.
            Uno solo (no admite listas ni "all").
          oneOf:
            - type: string
            - type: integer
              format: int64
          example: paola@ejemplo.com
        tipo:
          type: string
          enum: [monto, porcentaje]
          description: |
            "monto": descuenta una cantidad fija en pesos sobre el pedido.
            "porcentaje": descuenta un porcentaje del pedido.
        valor:
          type: number
          description: |
            Con tipo "monto": pesos, de 1 a 2000.
            Con tipo "porcentaje": entero de 1 a 30.
          example: 200
        nota:
          type: string
          maxLength: 120
          description: Texto corto que el profesional ve junto al cupón en su wallet.
          example: Por tu reseña del producto
        vence_en_dias:
          type: integer
          minimum: 1
          maximum: 365
          description: Días de vigencia desde hoy. Sin este campo, no vence.
        event_id:
          type: string
          maxLength: 120
          description: |
            Identificador único del evento. MANDARLO SIEMPRE: un reintento con el
            mismo `event_id` no crea un segundo cupón.
          example: resena_5521
        avisar:
          type: boolean
          default: true
          description: |
            Con false no se manda el aviso al teléfono (igual queda en la wallet). Útil si
            el agente prefiere redactar su propio mensaje con /api/webhooks/notificar.
        title:
          type: string
          maxLength: 100
          description: |
            Título del aviso. Va junto con `body`: los dos o ninguno. Sin ellos se usa el
            mensaje estándar. Admite los huecos `{codigo}`, `{descuento}`, `{valor}` y `{nota}`.
          example: ¡Felicidades! 🎉
        body:
          type: string
          maxLength: 300
          description: |
            Cuerpo del aviso. Se llena con el cupón recién creado, así el agente no necesita
            conocer el código antes: `{codigo}` es el código, `{descuento}` es por ejemplo
            "10% de descuento" o "$200 MXN de crédito", `{valor}` el número y `{nota}` la nota.
          example: Ganaste un {descuento} por tu reseña. Usa el código {codigo} en tu próxima compra.

    CuponCreado:
      type: object
      properties:
        ok:
          type: boolean
          const: true
        code:
          type: string
          description: El código del cupón.
          example: VH-K7M2PQ9X
        url:
          type: string
          format: uri
          description: Enlace que aplica el cupón al abrir la tienda.
        tipo:
          type: string
          enum: [monto, porcentaje]
        valor:
          type: number
        profesional:
          type: object
          properties:
            id:
              type: integer
              format: int64
            nombre:
              type: string
        exchange_id:
          description: Identificador del registro en la wallet del profesional.

    CuponDuplicado:
      type: object
      properties:
        ok:
          type: boolean
          const: true
        duplicado:
          type: boolean
          const: true
        code:
          type: string
        url:
          type: string
          format: uri
        mensaje:
          type: string

    Resumen:
      type: object
      description: Qué pasó con cada destinatario.
      properties:
        ok:
          type: boolean
          const: true
        destinatarios:
          type: integer
          description: Profesionales encontrados para ese `target`.
        enviados:
          type: integer
          description: Notificaciones que la plataforma de push aceptó.
        sin_token:
          type: integer
          description: |
            No tienen la app instalada. NO es un error: igual les queda la
            notificación en su panel dentro de la plataforma web.
        fallidos:
          type: integer
          description: La plataforma de push las rechazó.
        errores:
          type: array
          items:
            type: string
          description: |
            Motivos únicos de los rechazos. El más frecuente es
            "DeviceNotRegistered": el profesional desinstaló la app o cambió de
            teléfono. No es un problema de la integración y no hay que
            reintentar.

    Duplicado:
      type: object
      description: El `event_id` ya se había notificado; no se envió nada.
      properties:
        ok:
          type: boolean
          const: true
        duplicado:
          type: boolean
          const: true
        mensaje:
          type: string

    Error:
      type: object
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: string
          description: Qué falló, en texto legible.